+++ src/__playground__/index.ts
@@ -1,29 +1,10 @@
-import { worker } from './worker.js';
+import { SharedMap } from '@shared_map';
 
-const service = await worker.launchService();
+const map = new SharedMap({ foo: 'bar', test: 123 });
 
-try {
-    await service.call({ name: 'exit1' });
-} catch (error) {}
+await map.delete('foo');
 
-await new Promise((resolve) =>
-    setImmediate(async () => {
-        console.log(service.closed);
-        setTimeout(resolve, 2e3);
-    })
-);
+console.log(await map.get('foo'));
 
-await service.call({ name: 'exit1' });
-
-// const cluster = await worker.clusterize(5, {
-//     cluster: {
-//         autoRenew: false,
-//     },
-// });
-
-// console.log('calling');
-// const data = await cluster.use().call({ name: 'exit1' });
-// console.log(data, 'done');
-
-// console.log('foo');
-// console.log(cluster.activeServices);
+// @ts-ignore
+console.log(await map.get('xyz'));