+++ CHANGELOG.md
@@ -15,6 +15,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - `ServiceClusterOptions` and optional `autoRenew` option.
 - Accept exit codes in the `Service.close()` and `exceptionHandler.terminate()` functions.
+- `SharedMap.delete()` method.
+
+### Fixed
+
+- `SharedMap.get()` returning `null` for non-existent properties, but returning the string of `'null'` for existing properties with values set to `null`.
 
 ### Fixed